refactor: Add component js files
