+++ src/Square.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+export default (props) => {
+  return (
+    <button className="square" onClick={props.onClick}>
+      {props.value}
+    </button>
+  );
+}
